Shift DateAndTimeFormatter method from Event into Task
